Moved the RedundantPatternFound exception
The RedundantPatternFound exception exception is now found with all the other Asteroid exceptions in the asteroid_globals file.
