+++ code/asteroid_walk.py
@@ -1469,44 +1469,6 @@ dispatch_dict = {
 }
 
 ##############################################################################################
-# *** Part of the Redundant Pattern Detector ***
-#
-# This exception is used when a pattern has been identified as being 'useless',
-# or reundant. This exception is formatted to pack both offending patterns
-# information into a single formatted message to the user informing them of
-# where and what caused this error.
-##############################################################################################
-class RedundantPatternFound(Exception):
-    """Exception raised for detection of redundant pattern in function declaration.
-    Attributes:
-        pattern1 -- The pattern with the higher order of precedence (comparer).
-        line1    -- The line number location of pattern 1 in its file.
-        pattern2 -- The pattern with the lower order of precedence  (comparee).
-        line2    -- The line number location of pattern 2 in its file.
-        function -- The name of the function where the redundancy was found.
-        file     -- The name of the file where the function is from.
-    """
-    def __init__(self, pattern1, pattern2,function_name,location1,location2):
-
-        self.pattern1 = pattern1
-        self.line1 = str(location1[1] - 1) #patterns dont have line #, so we
-                                            #compensate here by using the first line
-                                            # of the patterns body, then moving back
-                                            # 1 line(minus 1)
-        self.pattern2 = pattern2
-        if (location2 != None):
-            self.line2 = str(location2[1] - 1)
-        if (location1 != None):
-            self.file = location1[0]
-        self.function = function_name
-        self.message = "Redundant Pattern Detected\n"
-        self.message += "\tFunction: " + self.function + " from file " + self.file
-        self.message += "\n\tPattern: " + term2string(self.pattern1) + " on line " + self.line1
-        self.message += "\n\twill consume all matches for"
-        self.message += "\n\tPattern: " + term2string(self.pattern2) + " on line " + self.line2
-        super().__init__(self.message)
-
-###########################################################################################
 # *** The Redundant Pattern Detector ***
 #
 # Evaluates the presence of redundant, or 'useless', pattern clauses in an Asteroid function:
@@ -1573,10 +1535,10 @@ def check_redundancy( body_list, f_name ):
             # the lower level of precedence useless/redundant by calling
             # on the unify function to evaluate the subsumption relationship
             # between the two patterns.
-            try:
-                unify( ptrn_l, ptrn_h , False )
-            except PatternMatchFailed:
-                pass
-            else:
+            try:                                #CHECK FOR CONFLICTION
+                unify( ptrn_l, ptrn_h , False ) 
+            except PatternMatchFailed:          #NO CONFLICTION
+                pass                            
+            else:                               #CONFLICTION
                 raise RedundantPatternFound( ptrn_h , ptrn_l , f_name, location_h, location_l )
 #######################################################################################
